Add all constants to folders
