+++ src/main/java/frc/robot/commands/paths/MoveToSingleSubstation.java
@@ -8,7 +8,7 @@ import com.pathplanner.lib.PathConstraints;
 import com.pathplanner.lib.PathPlanner;
 import com.pathplanner.lib.PathPoint;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.Constants.WaypointPositionConstants;
+import frc.robot.constants.WaypointPositionConstants;
 import frc.robot.Robot;
 import frc.robot.subsystems.drive.DriveBaseSubsystem;
 import frc.robot.subsystems.drive.GenerateTrajectory;
