+++ src/main/java/frc/robot/constants/ArmConstants.java
@@ -1,0 +1,13 @@
+package frc.robot.constants;
+
+public final class ArmConstants {
+
+  // testing purposes
+  public static final double mainArmPowerCoefficient = 0.5;
+  public static final double extendedArmPowerCoefficient = 0.3;
+
+  // setpoints for intaking and scoring (arbitrary values until we get arms to tune)
+  public static final double intakeSetpoint = 1000;
+  public static final double scoreSetpoint = 0;
+  public static final double resetSetpoint = 0;
+}
