+++ src/main/java/frc/robot/constants/ElevatorConstants.java
@@ -1,0 +1,26 @@
+package frc.robot.constants;
+
+import edu.wpi.first.math.trajectory.TrapezoidProfile;
+import edu.wpi.first.math.trajectory.TrapezoidProfile.Constraints;
+
+public final class ElevatorConstants {
+    // setpoints for scoring (arbitrary values until we get elevator to tune)
+    public static final double carriageMass = 8.381376;
+    public static final double drumRadius = 1.7;
+    public static final double elevatorGearing = 25;
+    public static final Constraints m_constraints = new TrapezoidProfile.Constraints(8.6, 4.3);
+    public static final double intakeSetpoint = 5;
+
+    public static enum NodeState {
+      GROUND(0),
+      SUBSTATION(20),
+      LOW(10),
+      HIGH(15);
+
+      public final double elevatorSetpoint;
+
+      private NodeState(double elevatorSetpoint) {
+        this.elevatorSetpoint = elevatorSetpoint;
+      }
+    }
+  }
