+++ src/main/java/frc/robot/constants/VisionConstants.java
@@ -1,0 +1,13 @@
+package frc.robot.constants;
+
+public final class VisionConstants {
+    public static final String name1 = "terima";
+    public static final String name2 = "teripaapa";
+    public static final double kTargetHeight = 1.071626; // meters
+    public static final double kCameraHeight = 0.5334; // meters
+    public static final double kCameraPitch = 42;
+    public static final double visionAmbiguityThreshold = 0.2;
+    public static final double focalLength = 2.9272781257541;
+    public static final int cameraResolutionWidth = 320; // pixels
+    public static final int cameraResolutionHeight = 240; // pixels
+}
