+++ src/main/java/frc/robot/subsystems/arm/MoveArmWithJoystick.java
@@ -2,7 +2,6 @@ package frc.robot.subsystems.arm;
 
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants.ArmConstants;
 
 public class MoveArmWithJoystick extends CommandBase {
 
@@ -28,13 +27,11 @@ public class MoveArmWithJoystick extends CommandBase {
   /** The execute allows us to control the robot using a joystick */
   @Override
   public void execute() {
-    if(joystick.getRightBumper()) {
+    if (joystick.getRightBumper()) {
       armSubsystem.setMainPower(0.1);
-    }
-    else if(joystick.getLeftBumper()) {
+    } else if (joystick.getLeftBumper()) {
       armSubsystem.setMainPower(-0.1);
-    }
-    else {
+    } else {
       armSubsystem.setMainPower(0);
     }
     // double mainPowerJoystick = ArmConstants.mainArmPowerCoefficient * joystick.getRightY();
