+++ src/main/java/frc/robot/subsystems/beambreak/BeamBreakSubsystem.java
@@ -6,14 +6,14 @@ package frc.robot.subsystems.beambreak;
 
 import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
+import frc.robot.constants.DeviceIDs;
 
 public class BeamBreakSubsystem extends SubsystemBase {
   private DigitalInput beamBreakReceiver;
 
   /** Creates a new BeamBreakSubsystem. */
   public BeamBreakSubsystem() {
-    beamBreakReceiver = new DigitalInput(Constants.SensorIds.beambreak.id);
+    beamBreakReceiver = new DigitalInput(DeviceIDs.SensorIds.beambreak.id);
   }
 
   @Override
