+++ src/main/java/frc/robot/subsystems/drive/ArcadeDrive.java
@@ -47,16 +47,16 @@ public class ArcadeDrive extends CommandBase {
     double xAxisSpeed = speedLimiter.calculate(joystickInputPower);
 
     double joystickInputPowerTurn = joystick.getRightX() * kTurn;
-    double zAxisRotation = joystickInputPowerTurn;
+    double zAxisRotation = -joystickInputPowerTurn;
     driveBaseSubsystem.drive(xAxisSpeed, zAxisRotation);
 
-    driveBaseSubsystem.coast();
+    // driveBaseSubsystem.coast();
 
-    double leftPower = xAxisSpeed + zAxisRotation;
-    double rightPower = xAxisSpeed - zAxisRotation;
+    // double leftPower = xAxisSpeed + zAxisRotation;
+    // double rightPower = xAxisSpeed - zAxisRotation;
 
-    driveBaseSubsystem.setLeftPower(leftPower);
-    driveBaseSubsystem.setRightPower(rightPower);
+    // driveBaseSubsystem.setLeftPower(leftPower);
+    // driveBaseSubsystem.setRightPower(rightPower);
   }
 
   @Override
