+++ src/main/java/frc/robot/subsystems/elevator/ElevatorToSetpoint.java
@@ -5,7 +5,7 @@
 package frc.robot.subsystems.elevator;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants.ElevatorConstants.NodeState;
+import frc.robot.constants.ElevatorConstants.NodeState;
 
 public class ElevatorToSetpoint extends CommandBase {
   private ElevatorSubsystem elevatorSubsystem;
@@ -36,5 +36,4 @@ public class ElevatorToSetpoint extends CommandBase {
     elevatorSubsystem.setPower(0);
     elevatorSubsystem.brake();
   }
-
 }
