+++ src/main/java/frc/robot/subsystems/gripper/RunGripper.java
@@ -5,7 +5,7 @@
 package frc.robot.subsystems.gripper;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants.WristConstants.GripperState;
+import frc.robot.constants.WristConstants.GripperState;
 
 public class RunGripper extends CommandBase {
